disable OS Login by default for GCP VMs (#708)
If [OS Login](https://cloud.google.com/compute/docs/oslogin) is enabled
in a GCP project, new VMs will be created with OS Login enabled by
default. When OS Login is enabled, `skyplane` is unable to establish an
SSH connection to the VM. Disabling OS Login (when allowed by the GCP
org) allows SSH connections via public key, and `skyplane` is able to
connect.

**Steps to reproduce OS Login issue:**
1. [Enable OS Login in a GCP
project](https://cloud.google.com/compute/docs/oslogin/set-up-oslogin#enable_os_login)
2. Run a Skyplane task which results in a GCP VM being provisioned
3. The task should fail with an error related to establishing an SSH
connection

So the provisioning succeeds, but attempting to connect to the instance
will fail. If you connect to the GCP VM directly (because it does not
get de-provisioned in this failure mode) and run `sudo journalctl -u
sshd` there will also be an error logged containing the message `Invalid
user skyplane`. What seems to be happening is that when OS Login is
enabled, SSH connections are rejected unless the user is known to OS
Login.

If OS Login is [required at the GCP organization
level](https://cloud.google.com/compute/docs/oslogin/manage-oslogin-in-an-org#set-org-policy)
then the provisioning process itself should fail. This seems fine
because it doesn't appear Skyplane can support provisioning and
connecting to VMs if OS Login is enabled, so the tool can consider
disabling OS Login a requirement and the error message seems
appropriate. If Skyplane can support connecting to a VM with OS Login
enabled then this approach may not be sufficient and could be misleading
in some cases.