Resolve redundant GCP logs in load_gcp_config (#786)
Intended to resolve #627

With my understanding, the redundancy was that "GCP credentials found"
was repeated twice. I adjusted the line in cli_init.py::load_gcp_config
as a result. Please let me know if there's a redundancy I'm not seeing.

All the other changes were due to running `autoflake --in-place
--remove-all-unused-imports --remove-unused-variables --recursive
skyplane` as per the contribution guidelines.